@@ -1,9 +1,9 @@
 {
   "name": "mat-select-filter",
-  "version": "2.3.5",
+  "version": "2.3.6",
   "peerDependencies": {
-    "@angular/common": "^7.1.0",
-    "@angular/core": "^7.1.0"
+    "@angular/common": "^8.1.3",
+    "@angular/core": "^8.1.3"
   },
   "main": "bundles/mat-select-filter.umd.js",
   "module": "fesm5/mat-select-filter.js",
